feat: fix department page header
