+++ package.json
@@ -24,6 +24,7 @@
     "react-dom": "18.2.0",
     "react-icons": "^4.7.1",
     "react-is": "^18.2.0",
+    "recoil": "^0.7.6",
     "styled-components": "^5.3.6",
     "styled-system": "^5.1.5",
     "typescript": "4.9.3"
