+++ src/atoms/EventsAtom.ts
@@ -1,0 +1,7 @@
+import { atom } from 'recoil';
+import { PresenceEvent } from 'types';
+
+export const eventListState = atom({
+  key: 'EventList',
+  default: [] as PresenceEvent[],
+});
