+++ src/components/index.ts
@@ -10,3 +10,4 @@ export * from './TextAndImage';
 export * from './Typography';
 export * from './VerticalLine';
 export * from './DescriptionCard';
+export * from './EventsLoader';
