+++ src/utils/constants.ts
@@ -1,0 +1,5 @@
+export const USU_ORGS = [
+  'Center for Student Involvement',
+  'Cross Cultural Centers',
+  'Recreation',
+];
